feat: mobile view flow
